fix(middleware): do not match tenant id
The middleware was copied from the vercel multi-tenant example which
includes a tenant id matcher. However, it appears the example has a bug
which matches paths such as `/a.b.c/xyz` which causes the locale not to
be rendered and sets the locale as `a.b.c` which ends up throwing an
error. Rather, the multi-tenancy works as expected without this Regex
because we use `NextResponse.rewrite` which proxies[1] the branded
request while preserving the original URL. In other words, currently a
request to `/code.org/en-US/engineering/all-the-things` works (when it
shouldn't), but the middleware using the rewrite proxy
`/en-US/engineering/all-the-things` also works because it proxies the
request internally.

This should fix some security robots throwing 500s in production.

[1] https://nextjs.org/docs/app/api-reference/functions/next-response#rewrite