fix: set keep alive greater than ELB timeout
The Node.js default keep alive is 5 seconds and the ELB default keep
alive is 60 seconds. ELB has a feature where it will preconnect to the
endpoint but has a race condition where if the endpoint is closed prior
(due to the 5 second timeout), the ELB will try to connect to a closed
endpoint. As documented in the Next.js server documentation[1], set the
timeout greater than the ELB value.

[1] https://nextjs.org/docs/app/api-reference/cli/next#configuring-a-timeout-for-downstream-proxies