fix: skip redirect test unless chrome
Skips the redirect test unless chrome because the logic is the same for
all browsers. On Safari, the test was acting flaky with the browser
unexpectedly closing and the behavior is more or less equivalent across all
browsers. This should also speed up the tests by skipping it on
firefox/webkit.