fix(404): return 404 when an api route is not found
When security scanners run through the /api/ routes, a 500 is generated
by the application because of a Next.js bug[1]. This bug causes the 404
error to be interpreted as a 500 which leads to error pollution in the
production logs.

To workaround the bug, a catch all route was added to catch unknown API
routes and explicitly set that the pages are 404. This causes the routes
to then show the sad bee page.

[1] https://github.com/vercel/next.js/issues/60477