feat(google): add site verification
This PR adds google site verifications that differentiate between
studio.code.org and code.org. For marketing sites, since the application
is multi-tenant, the meta header was chosen to allow for verification
based on runtime brand detection. In the future, we can switch to the
Domain TXT method to reduce page payload size if needed.