fix(canonical): all localized paths are canonical pages
Previously, the canonical tag was set on Contentful. However, canonical
pages (pages that are unique in content but de-duplicated if the content
is the same) are all our localized pages because they should have
localized content and therefore not be duplicates. This changes the
logic to automatically calculate the canonical path without having to
require content editors to define it.