Remove Temporary `enable-expire-other-sessions` DCDO Flag (#67328)
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/59763, which added some functionality to user's Account Settings page but gated it behind a DCDO flag so we could quickly turn it off if it revealed any problems.

The feature has been live for several weeks now, and we have (to my knowledge) no reported problems related to it. At this point I think it's safe to call this feature stable and remove the flag.