fix(cache): store cache-control on s3
This PR changes the cache-control mechanism for S3 assets from being set
on Cloudfront indiscriminatly via a response header override to being
stored in S3 as a Cache-Control metadata header.

Previously, the response header always added `Cache-Control` regardless
of the origin response code which leads to circumstances where a 4xx or
5xx is cached (if Cloudfront caches that status code).

Now, S3 will be responsible for setting `Cache-Control`, for which it
will only respond with `Cache-Control` if the file was successfuly
retrieved (as it will be in that file's metadata).

This was tested on `csforall.marketing-sites.dev-code.org`.