fix(prettier): do not prettier transient files
When a concurrent build and lint step occurs, if the build is still
ongoing, prettier may pick up and attempt to fix the transient tsup
build file. This causes a build failure if the transient file was picked
up but could not later be read for processing. Since these are transient
files, there is no need to include them in git nor prettier them.