fix(cache): add cache behavior for images
This PR adds a cache behavior for `/_next/image*` on the Pegasus
distribution allowing cache variance on the `Accept` header to determine
whether a client can accept webp or avif image formats. Currently, this
header is not passed causing the application to mistakenly believe that
the client does not support webp or avif causing images to be sent out
as jpegs. Unfortunately, jpeg does not support alpha transparency and
therefore causes a black background to appear where the alpha
transparency is. To fix this, ensure that the accept header is passed
down. This fixes it for most clients, all the way down to chrome 65. For
older clients, they will see a degraded experience with the black
background as mentioned before. However, this can be fixed (if needed)
in the future by adjusting the application.