feat(Image): use next image for image
This PR updates the Image component to use Next.js dynamic image
component. This offloads images from Contentful's CDN to our self-hosted
one. To accomplish this, the asset id is retrieved from the image url[1]
which is fed into the `useInMemoryEntities` helper tool to retrieve the
entire Contentful data structure for the width and height (without
needing to add it to the definition itself). This then allows Next.js to
reserve image space to eliminate content layout shifting (because the
height and width is fixed). Currently, images are loaded without a fixed
height/width resulting in a CLS score reduction.

In the case we are unable to get the height and width, the regular
`<img>` element will be used instead.

[1] https://www.contentful.com/developers/docs/concepts/links/